# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cuba-api",
        "description": "Api para consultar la division administrativa de Cuba",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://cuba-api.onrender.com/api/v1/"
        }
    ],
    "paths": {
...
        "/provincias/{id}": {
            "get": {
                "tags": [
                    "Provincias"
                ],
                "summary": "Obtener Provincia",
                "description": "Obtiene todos los datos de todas las provincias con su relacion con los municipios y sus datos",
                "operationId": "4ec62e2f5241b4a53cafd22acde9222a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El id de la provincia",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "maximum": 16,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Artemisa"
                                                    },
                                                    "acronimo": {
                                                        "type": "string",
                                                        "example": "ART"
                                                    },
                                                    "es_capital": {
                                                        "type": "boolean",
                                                        "example": "false"
                                                    },
                                                    "municipios": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id_municipio": {
                                                                    "type": "number",
                                                                    "example": "1"
                                                                },
                                                                "nombre_municipio": {
                                                                    "type": "string",
                                                                    "example": "Alquizar"
                                                                },
                                                                "es_cabecera_provincial": {
                                                                    "type": "boolean",
                                                                    "example": "false"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "msg": {
                                                        "type": "string",
                                                        "example": "La provincia con ese id no existe"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}